Fix TagTrait
