+++ web/modules/custom/server_general/src/EntityViewBuilder/NodeViewBuilderAbstract.php
@@ -4,8 +4,8 @@ namespace Drupal\server_general\EntityViewBuilder;
 
 use Drupal\pluggable_entity_view_builder\EntityViewBuilderPluginAbstract;
 use Drupal\server_general\ProcessedTextBuilderTrait;
+use Drupal\server_general\TagTrait;
 use Drupal\server_general\ThemeTrait\ElementWrapTrait;
-use Drupal\server_general\ThemeTrait\TagTrait;
 
 /**
  * An abstract class for Node View Builders classes.
